+++ WebStore/Views/Emp/EmployeeDetails.cshtml
@@ -18,10 +18,10 @@
 
     <p>Офис: @Model.Office</p>
     <p>Хобби: @Model.Hobby</p>
-    <p>Дата рождения: @Model.Birthdate</p>
+    <p>Дата рождения: @Model.Birthdate.ToShortDateString()</p>
 
     <hr />
-    <a href="/Home/Index">К списку сотрудников</a>
+    <a asp-controller="Emp" asp-action="Index">К списку сотрудников</a>
 
 </body>
 </html>
