+++ WebStore/Views/Emp/Index.cshtml
@@ -1,0 +1,47 @@
+﻿    @model IEnumerable<WebStore.Models.Employee>
+
+    <!DOCTYPE html>
+
+    <html>
+    <head>
+        <meta name="viewport" content="witdh=device-witdh" />
+        <title>Первое приложение @ViewBag.Title</title>
+    </head>
+
+    <body>
+        <h3>Сотрудники!</h3>
+
+        <table>
+            <thead>
+                <tr>
+                    <th>id</th>
+                    <th>Фамилия</th>
+                    <th>Имя</th>
+                    <th>Отчество</th>
+                    <th>Возраст</th>
+                    <th>Офис</th>
+                    <th>Хобби</th>
+                    <th>Дата рождения</th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                @foreach (var employee in Model)
+                {
+                    <tr>
+                        <td>@employee.Id</td>
+                        <td>@employee.Surname</td>
+                        <td>@employee.FirstName</td>
+                        <td>@employee.Patronymic</td>
+                        <td>@employee.Age</td>
+                        <td>@employee.Office</td>
+                        <td>@employee.Hobby</td>
+                        <td>@employee.Birthdate.ToShortDateString()</td>
+                        <td><a asp-controller="Emp" asp-action="EmployeeDetails" asp-route-id="@employee.Id">Детали</a></td>
+                    </tr>
+                }
+            </tbody>
+        </table>
+
+    </body>
+</html>
