+++ WebStore/Views/Home/Index.cshtml
@@ -1,6 +1,4 @@
-﻿@model IEnumerable<WebStore.Models.Employee>
-
-<!DOCTYPE html>
+﻿<!DOCTYPE html>
 
 <html>
 <head>
@@ -11,37 +9,7 @@
 <body>
     <h3>Сотрудники!</h3>
 
-    <table>
-        <thead>
-            <tr>
-                <th>id</th>
-                <th>Фамилия</th>
-                <th>Имя</th>
-                <th>Отчество</th>
-                <th>Возраст</th>
-                <th>Офис</th>
-                <th>Хобби</th>
-                <th>Дата рождения</th>
-                <th></th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var employee in Model)
-            {
-            <tr>
-                <td>@employee.Id</td>
-                <td>@employee.Surname</td>
-                <td>@employee.FirstName</td>
-                <td>@employee.Patronymic</td>
-                <td>@employee.Age</td>
-                <td>@employee.Office</td>
-                <td>@employee.Hobby</td>
-                <td>@employee.Birthdate</td>
-                <td><a href="/Home/EmployeeDetails/@employee.Id">Инфо.</a></td>
-            </tr>
-            }
-        </tbody>
-    </table>
+    <a asp-controller="Emp" asp-action="Index">Список</a>
 
 </body>
 </html>
