+++ WebStore/Views/_ViewImports.cshtml
@@ -1,0 +1,2 @@
+﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, WebStore
