+++ WebStore/appsettings.json
@@ -7,5 +7,5 @@
     }
   },
   "AllowedHosts": "*",
-  "CustomGreetings" : "Hello from ASP.NET Core App 3.1!!! -----------------!!!!!" 
+  "CustomGreetings" : "Hello from ASP.NET Core App 3.1!!!!!!" 
 }
